dental_clinic/requirements.txt

Babel==2.15.0

django-phonenumber-field==8.0.0
django-picklefield==3.2
django-q2==1.6.2

phonenumbers==8.13.42
pillow==10.4.0

